% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{cortical.plot}
\alias{cortical.plot}
\title{Cortical plot}
\usage{
cortical.plot(dataset, type = c("left", "right", "top"),
  show.sections = TRUE, labels = TRUE, ml.cut = 2, dv.cut = -2.8,
  col = NULL, cex = 0.5, pch = 16, pt.bg = NULL)
}
\arguments{
\item{dataset}{a dataset frame obtained by get.cell.ids() or by inspect.registration().}

\item{type}{character vector, what type of plots should be made. Default is c('left', 'right', 'top').}

\item{show.sections}{boolean, if lines should be drawn for each unique AP section. Default is TRUE.}

\item{labels}{if labels should be plotted with acronyms for each cortical region.}

\item{ml.cut}{numeric, value medio-lateral plane where everything more medial than ml.cut will be excluded. Default is 2.0 mm.}

\item{dv.cut}{numeric, value dorso-ventral plane where everything above dv.cut will be plotted. Default is -2.8 mm.}

\item{col}{color code or name.}

\item{pch}{plotting ‘character’, i.e., symbol to use. This can either be a single character or an integer code for one of a set of graphics symbols.}

\item{pt.bg}{background (fill) color for the open plot symbols given by pch = 21:25.}
}
\description{
Plots schematic brain(s) of cells in cortex from different views.
}
\examples{
ccortical.plot(dataset, type = c('left', 'right', 'top') )
}
