% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{cortical.sideview}
\alias{cortical.sideview}
\title{Cortical sideview plot}
\usage{
cortical.sideview(dataset, right.hemisphere = TRUE,
  show.sections = TRUE, labels = TRUE, col = NULL, pch = 16,
  pt.bg = NULL)
}
\arguments{
\item{dataset}{a dataset frame obtained by get.cell.ids() or by inspect.registration().}

\item{right.hemisphere}{boolean, if true then right side of the hemisphere will be plotted, if false left hemisphere will be plotted.}

\item{show.sections}{boolean, if lines should be drawn for each unique AP section. Default is TRUE.}

\item{labels}{if labels should be plotted with acronyms for each cortical region.}

\item{col}{color code or name.}

\item{pch}{plotting ‘character’, i.e., symbol to use. This can either be a single character or an integer code for one of a set of graphics symbols.}

\item{pt.bg}{background (fill) color for the open plot symbols given by pch = 21:25.}
}
\description{
Plots a schematic brain from the side with cortical regions outlined.
}
\examples{
cortical.sideview(dataset, right.hemisphere = FALSE, labels = FALSE)
}
