% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{cortical.topview}
\alias{cortical.topview}
\title{Cortical topview plot}
\usage{
cortical.topview(dataset, dv.cut = -2.8, show.sections = TRUE,
  labels = TRUE, col = NULL, pch = 16, cex = 0.5, pt.bg = NULL)
}
\arguments{
\item{dataset}{a dataset frame obtained by get.cell.ids() or by inspect.registration().}

\item{dv.cut}{numeric, value dorso-ventral plane where everything above dv.cut will be plotted. Default is -2.8 mm.}

\item{show.sections}{boolean, if lines should be drawn for each unique AP section. Default is TRUE.}

\item{labels}{if labels should be plotted with acronyms for each cortical region.}

\item{col}{color code or name.}

\item{pch}{plotting ‘character’, i.e., symbol to use. This can either be a single character or an integer code for one of a set of graphics symbols.}

\item{pt.bg}{background (fill) color for the open plot symbols given by pch = 21:25.}
}
\description{
Plots a schematic brain from the top showing only parts that are easily accessible with cranial windows without removing parts of the temporalis muscle.
}
\examples{
cortical.topview(dataset, right.hemisphere = FALSE, labels = FALSE)
}
