% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{data.summary}
\alias{data.summary}
\title{Summary mean and dispersion}
\usage{
data.summary(data, varname, groupnames)
}
\arguments{
\item{varname}{the name of a column containing the variable to be summariezed}

\item{a}{data frame}

\item{groupnamesvector}{of column names to be used as grouping variables}
}
\description{
Function to calculate the mean and the dispersion for each group
}
\examples{
df <- data.summary(data, varname="cell.count", groupnames=c("genotype", "dose"))
}
