% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{draw.blobs}
\alias{draw.blobs}
\title{Draw fluorescent blobs}
\usage{
draw.blobs(input, col = rgb(1, 0, 0, 0.5), border = "darkred")
}
\arguments{
\item{input}{a segmentation object obtained by segment() with get.contour = TRUE.}
}
\description{
Draws fluorescent blobs.
}
\examples{
plot.registration(regi)
draw.blobs(seg, col = rgb(0,1,0.1, alpha = 0.5), border ='green3')
}
