% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getgridcoordiantes.R
\name{find.dim}
\alias{find.dim}
\title{Find dimensions}
\usage{
find.dim(x)
}
\arguments{
\item{x}{an integer value which give sthe number of tiles.}
}
\description{
Find dimensions of stitched image from number of tiles.
}
\examples{
find.dim(63)
}
\keyword{grid}

