% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatfieldcorrection.R
\name{flat.field.correction}
\alias{flat.field.correction}
\title{Flat-field correction}
\usage{
flat.field.correction(input, output.folder = "../",
  output.prefix = "FFC", kernel = 301, show.image = FALSE,
  gain.image.name = "gain_{output.folder}.tif", verbose = TRUE)
}
\arguments{
\item{input}{input a character vector consisting of the full path name to 16-bit raw tif image files. For microscopy the most common systematic error is vignetting where dark shades are stronger among the pixels on the edges.}

\item{output.prefix}{the prefix for the generated image tiles, default is FFC_.}

\item{kernel}{smoothing kernel for the Gaussian smothing of the gain image, higher means more bluring, the number needs to be an odd number default is 301.}

\item{show.image}{a boolean value, if true the stitched image will be displayed in a display window. Default is false.}

\item{gain.image.name}{the name of the gain image to be generated default is adding the prefix gain_ to the folder name.}

\item{verbose}{boolean value. If true diagnostic output is written to the R console. Deafult is true.}

\item{outputfolder}{name of output folder to save the image in. Default is ../ which means a directory will be created (if not allready there) in the parent directory to where the tiles are placed. Alternative is either ./ which will create output folder in the current directory in which the tiles are placed. The user might also provide the full system path to an already existing directory.}
}
\description{
An image sensor of a camera is composed of a two dimensional array of light sensitive detectors or pixels. The sesnor array is #'mechanically quite stable with the pixels retaining a rigidly fixed geometric relationship. Each pixel within the array, however, #'has its own unique light sensitivity characteristics. As these characteristics affect camera performance, they must be removed #'through calibration. The process by which a camera is calibrated is known as "Flat Fielding" or "Shading Correction".
}
\examples{
#folder where image tiles are stored
images<-get.images('/Volumes/microscope/animal001/slide001/section001')
#stitch images
flat.field.correction(images)
}
