% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{get.atlas.image}
\alias{get.atlas.image}
\title{Get atlas binary image}
\usage{
get.atlas.image(coordinate, directory = "TEMPORARY", width = 456,
  plane = "coronal", plate = FALSE, right.hemisphere = NULL,
  close.image = TRUE, save.image = TRUE)
}
\arguments{
\item{coordinate}{matching coordinates in the main plane to be registered to (in millimeters).}

\item{directory}{either an absolute path or the default TEMPORARY which will save the tif output to .}

\item{width}{the pixel width to save as TIFF, default is 456px, but 11300 is high res.}

\item{plate}{if true coordinate indicates plate number out of 132 plates.}

\item{verbose}{boolean value. If true diagnostic output is written to the R console. Deafult is true.}
}
\description{
Get atlas binary image
}
\examples{
#path to image
image<-'/Volumes/microscope/animal001/slide001/section001.tif'
#register the image
registration(image, AP=1.05, brain.threshold=220)
}
