% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{get.cortex.layer}
\alias{get.cortex.layer}
\title{Get index over cells in specific cortical layers}
\usage{
get.cortex.layer(dataset = dataset, layer = c(1:6))
}
\arguments{
\item{dataset}{a dataset frame obtained by get.cell.ids() or by inspect.registration().}

\item{layer}{integer vector with unique layers of interest c(1:6) will give you all layers, default is c(1:6).}
}
\description{
Returns a two-dimensional data frame with the first column, layer, indicating which layer the cell is in and the second column, index, indicating which row in the original data frame the cell can be found.
}
