% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgridcoordiantes.R
\name{get.grid.coordinates}
\alias{get.grid.coordinates}
\title{Get Grid Coordinates}
\usage{
get.grid.coordinates(input, tilesize, overlap = 0.1, rotate = 0,
  plotgrid = FALSE)
}
\arguments{
\item{input}{vector of files, can be eithe rnumeric or character only importance is that length(input) is defined.}

\item{tilesize}{a single integer which givesthe size of the side of the image tile in pixels (assumes square).}

\item{overlap}{overlap expressed either as proportion (if overlap<1) of in pixels (if overlap>1).}
}
\description{
Given a list of file names as input or just a vector this function retrieves the coordinates for each tile as well as the overlaps. It returns a list object with two sublists; width is the image width in pixels, height is the image height in pixels, grid.layout is a data frame where each row is an image and grid.layout$top, grid.layout$bottom, grid.layout$left, grid.layout$right, indicates the position if any at the boundary (0 or 1). grid.layout$x0, grid.layout$x1, grid.layout$y0, grid.layout$y1 gives the coordinates of the tile in the stitched image. The object overlaps is a list with three data frames; overlaps$horizontal give sthe overlap in the horizontal direction of the image plane (right corner of each tile), overlaps$vertical give sthe bottom corner overlap, and overlaps$small gives the small intersection (normally perfectly square) between horizontal and vertical overlap.
}
\examples{
grid.coordinates<-get.grid.coordinates(1:63, 2048, plotgrid=F)
}
\keyword{grid}
