% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getimagesfromdirectory.R
\name{get.images}
\alias{get.images}
\title{Get images from directory}
\usage{
get.images(x, type = "tif")
}
\arguments{
\item{input}{a list of file names.}
}
\description{
Find dimensions of stitched image from number of tiles.
}
\examples{
get.images('/Volumes/Seagate Slim Drive/D159/FFC_section001_1_ownFFC')
}
\keyword{grid}
