% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{get.pixel.intensity}
\alias{get.pixel.intensity}
\title{Get pixel intensity}
\usage{
get.pixel.intensity(input, x, y, type = "SNR", roi = 9,
  background = 40)
}
\arguments{
\item{input}{file path to a valid monochrome 16-bit TIFF.}

\item{x}{pixel x coordinate(s).}

\item{y}{pixel y coordinate(s).}

\item{type}{what type of intensity value default is 'SNR' set to 'intensity' to get raw pixel intensities.}

\item{roi}{.}

\item{background}{.}
}
\description{
This function extracts the pixel intensity for a given set of x and y coordinates.
Returns a numeric vector of values.
}
\examples{
snr <- get.pixel.intensity(image, dataset$x, dataset$y, type='SNR', roi=10, background=40)
intensity <- get.pixel.intensity(image, dataset$x, dataset$y, type='intensity', roi=10)
}
\keyword{intensity}
\keyword{pixel}
