% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes.R
\name{get.probe.paths}
\alias{get.probe.paths}
\title{Get probe paths}
\usage{
get.probe.paths(
  dataset,
  id = NULL,
  surface.point = 0,
  cluster = FALSE,
  axis = c("ML", "DV", "AP")
)
}
\arguments{
\item{dataset}{a data frame generated by getcell.ids() or inspect.registration().}
}
\value{
output is returned in the form of probe.paths class.
}
\description{
This function uses a dataset data frame and returns probe paths.
Returns a probe.paths class.
}
\examples{
seg<-segment(FFC_filename)
plot(seg$soma$x, seg$soma$y, ylim=rev(range(seg$soma$y)), asp=1, pch=16)
#invoke text editor to extract filter for copy paste into your script for reprodicible results.
edit(seg$filter)
}
\keyword{probes}
