% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{get.region.volume}
\alias{get.region.volume}
\title{Get region volume}
\usage{
get.region.volume(regions, bilateral = FALSE)
}
\arguments{
\item{regions}{vector of either acronyms, names or integer ids of regions that you want volume of.}

\item{bilateral}{boolean. If volume should be calculated for a single hemisphere or both. Default is TRUE.}
}
\description{
Providing an acronym, id or name this function returns the volume in cubic millimeters of the region. Option to only output for single hemisphere.
Returns a numeric vector of values in cubic millimeters.
}
\examples{
volume <- get.region.volume(dataset$acronym, bilteral = FALSE)
}
