% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatewebfiles.R
\name{makewebmap}
\alias{makewebmap}
\title{Generate a web-map from a sample section}
\usage{
makewebmap(img, filter, registration = NULL, dataset = NULL,
  folder.name = NULL, scale = 0.64, bregmaX = 0, bregmaY = 0,
  fluorophore = "Rabies-EGFP", combine = NULL, enable.drawing = TRUE,
  write.tiles = TRUE, verbose = FALSE)
}
\arguments{
\item{img}{input a character vector consisting of the full path name to 16-bit raw tif image files.}

\item{verbose}{boolean value. If true diagnostic output is written to the R console. Deafult is true.}

\item{alpha}{filename of the stitched output, default is the directory where the tiles are situated with the added prefix stitched_}

\item{beta}{type of motorized microscope stage. Will define the order of acquisition. Arguments are either row.by.row or default snake.by.row.}

\item{cells}{the order each tile is acquired. Default is left.&.up which starts from bottom left corner and moves up. Other options include right.&.up, left.&.down and right.&.down.}

\item{outputfolder}{name of output folder to save the image in.}
}
\description{
Generates multiple image tiles in JPEG format.
}
\examples{
#folder where image tiles are stored
makewebmap('/Users/danielfurth/Documents/mouse_atlas/cropped_D1_59_s4.lif_TileScan_006_Merging001_ch00.tif', alpha=60, beta=10, verbose=TRUE)
#stitch images
stitch(images, type = 'snake.by.row', order = 'left.&.up', tilesize=2048, overlap=0.1, show.image=TRUE)
}
