% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{normalize.volume}
\alias{normalize.volume}
\title{Normalize volume}
\usage{
normalize.volume(dataset, bilateral = TRUE, aggregate = NULL)
}
\arguments{
\item{dataset}{a dataset frame obtained by get.cell.ids() or by inspect.registration().}

\item{bilateral}{boolean, if true then cells will be normalized by total volume regardless of hemisphere.}

\item{aggregate}{compute volume for subregins based on parent region. Add which parent region you want to aggregate.}
}
\description{
Returns a count table of cell counts per region normalized by cubic millimeters.
}
\examples{
densities<-normalize.volume(dataset, aggregate = c('MO', 'SSp'))
}
