% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{nuclear.segment}
\alias{nuclear.segment}
\title{Get contour}
\usage{
nuclear.segment(input, thresh = 0.1, kernel = 3, iter = 20,
  threshold = "otsu", invert = FALSE, get.largest.object = TRUE,
  num.nested.objects = 2, blur = 0, watershed = FALSE,
  resize = 0.25, display = TRUE, verbose = TRUE,
  savefilename = FALSE)
}
\arguments{
\item{input}{input a character vector consisting of the full path name to 16-bit raw tif image files.}

\item{verbose}{boolean value. If true diagnostic output is written to the R console. Deafult is true.}

\item{coordinate}{matching coordinates in the main plane to be registered to (in millimeters).}

\item{plane}{the main plane to register too: "coronal", "sagital", "".}

\item{brain.threshold}{a integer value, which determien sthe segmentation of the brain slice.}
}
\description{
Get the contours of a binary segmented image
}
\examples{
#path to image
image<-'/Volumes/microscope/animal001/slide001/section001.tif'
#register the image
registration(image, AP=1.05, brain.threshold=220)
}
