% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{pax.to.allen}
\alias{pax.to.allen}
\title{Get Allen 25 micron voxel coordinates from stereotactic}
\usage{
pax.to.allen(paxinos)
}
\arguments{
\item{paxinos}{AP form bregma.}
}
\description{
Returns integer of anterior-posterior voxel index.
}
