% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{reflect.cells}
\alias{reflect.cells}
\title{Get multi channel}
\usage{
reflect.cells(dataset)
}
\arguments{
\item{dataset}{a data frame generated by inspect.registration() or get.cell.ids().}

\item{seg}{a segmentation object.}

\item{images}{a character vector with file paths for images.}

\item{images}{a character vector with file paths for images.}

\item{right.hemisphere}{boolean, if TRUE then cells will be flipped to right side.}
}
\description{
Gets intensity or signal-to-noise measurement for segmented features in multiple channels.
}
\examples{
dataset<-reflect.cells(dataset)
Flip cells to the ipsilateral side

Flips the cells to the right hemisphere.
dataset<-reflect.cells(dataset)
}
