% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{registration}
\alias{registration}
\title{Register}
\usage{
registration(input, coordinate = NULL, plane = "coronal",
  right.hemisphere = NULL, interpolation = "tps", intrp.param = NULL,
  brain.threshold = 200, blurring = c(4, 15), pixel.resolution = 0.64,
  resize = (1/8)/4, correspondance = NULL, resolutionLevel = c(4, 2),
  num.nested.objects = 0, display = TRUE, plateimage = FALSE,
  forward.warp = FALSE, filter = NULL, output.folder = "../",
  batch.mode = FALSE, verbose = TRUE)
}
\arguments{
\item{input}{input a character vector consisting of the full path name to 16-bit raw tif image files.}

\item{coordinate}{matching coordinates in the main plane to be registered to (in millimeters).}

\item{plane}{the main plane to register too: "coronal", "sagital", "".}

\item{brain.threshold}{a integer value, which determien sthe segmentation of the brain slice.}

\item{verbose}{boolean value. If true diagnostic output is written to the R console. Deafult is true.}
}
\description{
An image sensor of a camera is composed of a two dimensional array of light sensitive detectors or pixels. The sesnor array is #'mechanically quite stable with the pixels retaining a rigidly fixed geometric relationship. Each pixel within the array, however, #'has its own unique light sensitivity characteristics. As these characteristics affect camera performance, they must be removed #'through calibration. The process by which a camera is calibrated is known as "Flat Fielding" or "Shading Correction".
}
\examples{
#path to image
image<-'/Volumes/microscope/animal001/slide001/section001.tif'
#register the image
registration(image, AP=1.05, brain.threshold=220)
}
