% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{registration2}
\alias{registration2}
\title{Register section to atlas}
\usage{
registration2(
  input,
  coordinate = NULL,
  filter = myfilter,
  interpolation = "tps",
  intrp.param = NULL,
  output.folder = "../",
  forward.warp = FALSE,
  display = TRUE,
  tissue.threshold = 200,
  blurring = c(4, 4),
  pixel.resolution = 0.64,
  resize = (1/8)/4,
  correspondance = NULL,
  resolutionLevel = c(4, 2),
  num.nested.objects = 0,
  batch.mode = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{input}{input a character vector consisting of the full path name to 16-bit raw tif image files.}

\item{coordinate}{matching coordinates in the main plane to be registered to (in millimeters).}

\item{verbose}{boolean value. If true diagnostic output is written to the R console. Deafult is true.}

\item{plane}{the main plane to register too: "coronal", "sagital", "".}

\item{brain.threshold}{a integer value, which determien sthe segmentation of the brain slice.}
}
\description{
This function performs registration using point-based algorithms either Thin-Plate B-Splines (TPS) or Coherent Point Drift (CPD).
}
\examples{
#path to image
image<-'/Volumes/microscope/animal001/slide001/section001.tif'
#register the image
regi<-registration(image, coordinate = 1.05, filter = seg$filter)
}
