% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staoseq.R
\name{remove.spot.outliers}
\alias{remove.spot.outliers}
\title{Remove spot outliers}
\usage{
remove.spot.outliers(segmentation, width = 26000, height)
}
\arguments{
\item{cells}{to be of H&E nucleu segmentation list obtained from segment()}

\item{array.dim}{the dimensions of the ST array. Default is c(33, 35)}

\item{tol.dist}{distance that is tolerated in pixels between the center of two spots. Default is 200 px.}
}
\value{
spot.id two column matrix giving the identity to each spot in row and column coordinates.

missing two column matrix giving the identity to each spot that couldn't be detected.
}
\description{
This takes segmented cell nuclei from H&E staining and spot contours form Cy3 ST array 
and returns a parent vector for each cell nuclei.
}
\examples{
prune.spots(seg)
}
