% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{roi.cell.count}
\alias{roi.cell.count}
\title{Get cell counts for a specific Region of Interest (ROI).}
\usage{
roi.cell.count(dataset, rois = c("MO", "TH"))
}
\arguments{
\item{dataset}{a data frame generated by inspect.registration() or get.cell.ids().}

\item{rois}{a character vector with acronyms for brain regions you want.}
}
\description{
Get cell counts for a specific Region of Interest (ROI) in tidy format.
}
\examples{
#get cell position and identity
dataset<-get.cell.ids(regi, seg)
#extract roi counts
data<-roi.cell.count(dataset, rois=c('MO','TH'))
}
