% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment}
\alias{segment}
\title{Segment image using multiple thresholds}
\usage{
segment(input, numthresh = 8, downsample = 0.25, filter = NULL,
  post = NULL, pre = NULL, get.contour = FALSE, display = TRUE)
}
\arguments{
\item{input}{input a character vector consisting of the full path name to 16-bit raw tif image files.}

\item{numthresh}{numbe rof binary thresholds to use (default is 8).}
}
\description{
Segment brain section.
}
\examples{
#folder where image tiles are stored
images<-get.images('/Volumes/microscope/animal001/slide001/section001')
#stitch images
segment(system.file('sample_tiles/rabiesEGFP.tif', package='wholebrain')) 
}
